# Lazily deserialize `StoreIndexElement`s `content`