--- a/versioned/storage/common/src/main/java/org/projectnessie/versioned/storage/common/persist/ObjId.java
+++ b/versioned/storage/common/src/main/java/org/projectnessie/versioned/storage/common/persist/ObjId.java
@@ -161,6 +161,11 @@ public abstract class ObjId {
     return fromBytes(len, bytes);
   }
 
+  public static void skipObjId(@Nonnull @jakarta.annotation.Nonnull ByteBuffer bytes) {
+    int len = readVarInt(bytes);
+    bytes.position(bytes.position() + len);
+  }
+
   private static ObjId fromBytes(int len, ByteBuffer bytes) {
     switch (len) {
       case 0: