--- a/versioned/storage/common/src/main/java/org/projectnessie/versioned/storage/common/util/Ser.java
+++ b/versioned/storage/common/src/main/java/org/projectnessie/versioned/storage/common/util/Ser.java
@@ -64,4 +64,13 @@ public final class Ser {
     }
     return r;
   }
+
+  public static void skipVarInt(ByteBuffer b) {
+    for (; ; ) {
+      int v = b.get() & 0xff;
+      if ((v & 0x80) == 0) {
+        break;
+      }
+    }
+  }
 }