The proposal described in #7061 _could_ bring a quite significant reduction in CPU and heap pressure, however re-serialization would add a lot more heap pressure.

This change introduces lazily deserializion of the _content_ of a `StoreIndexElement`. This already gives some nice improvement over the existing code to eagerly deserialize the _content_, because that is often not needed.

Below are the JMH results for an index with 156250 elements and a serialized size of 9228901 bytes. Heap pressure is reduced by about 25% for deserialzation (roughly 10M in this configuration).

Before this change (against the 1st commit in this PR):
```
Benchmark                                                                (deterministic)  (foldersPerLevel)  (namespaceLevels)  (tablesPerNamespace)  Mode  Cnt         Score      Error   Units
RealisticKeyIndexImplBench.deserialize                                              true                  5                  5                    50  avgt    3     14844.961 ± 2460.785   us/op
RealisticKeyIndexImplBench.deserialize:·gc.alloc.rate                               true                  5                  5                    50  avgt    3      2405.123 ±  399.027  MB/sec
RealisticKeyIndexImplBench.deserialize:·gc.alloc.rate.norm                          true                  5                  5                    50  avgt    3  37443286.030 ±    0.943    B/op
RealisticKeyIndexImplBench.deserialize:·gc.count                                    true                  5                  5                    50  avgt    3         4.000             counts
RealisticKeyIndexImplBench.deserialize:·gc.time                                     true                  5                  5                    50  avgt    3         8.000                 ms
RealisticKeyIndexImplBench.deserializeAdd                                           true                  5                  5                    50  avgt    3     11384.819 ± 5278.655   us/op
RealisticKeyIndexImplBench.deserializeAdd:·gc.alloc.rate                            true                  5                  5                    50  avgt    3      3216.608 ± 1510.121  MB/sec
RealisticKeyIndexImplBench.deserializeAdd:·gc.alloc.rate.norm                       true                  5                  5                    50  avgt    3  38388292.621 ±    2.171    B/op
RealisticKeyIndexImplBench.deserializeAdd:·gc.count                                 true                  5                  5                    50  avgt    3         5.000             counts
RealisticKeyIndexImplBench.deserializeAdd:·gc.time                                  true                  5                  5                    50  avgt    3         6.000                 ms
RealisticKeyIndexImplBench.deserializeAddSerialize                                  true                  5                  5                    50  avgt    3     23515.950 ± 9560.517   us/op
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.alloc.rate                   true                  5                  5                    50  avgt    3      1931.861 ±  774.923  MB/sec
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.alloc.rate.norm              true                  5                  5                    50  avgt    3  47627313.674 ±   12.676    B/op
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.count                        true                  5                  5                    50  avgt    3         3.000             counts
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.time                         true                  5                  5                    50  avgt    3         5.000                 ms
RealisticKeyIndexImplBench.deserializeGetRandomKey                                  true                  5                  5                    50  avgt    3     10915.480 ± 1236.127   us/op
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.alloc.rate                   true                  5                  5                    50  avgt    3      3270.873 ±  369.704  MB/sec
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.alloc.rate.norm              true                  5                  5                    50  avgt    3  37443308.451 ±    0.513    B/op
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.count                        true                  5                  5                    50  avgt    3         5.000             counts
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.time                         true                  5                  5                    50  avgt    3         9.000                 ms
RealisticKeyIndexImplBench.deserializeIterate250                                    true                  5                  5                    50  avgt    3     10857.502 ± 7434.119   us/op
RealisticKeyIndexImplBench.deserializeIterate250:·gc.alloc.rate                     true                  5                  5                    50  avgt    3      3291.342 ± 2206.207  MB/sec
RealisticKeyIndexImplBench.deserializeIterate250:·gc.alloc.rate.norm                true                  5                  5                    50  avgt    3  37443316.526 ±    6.609    B/op
RealisticKeyIndexImplBench.deserializeIterate250:·gc.count                          true                  5                  5                    50  avgt    3         5.000             counts
RealisticKeyIndexImplBench.deserializeIterate250:·gc.time                           true                  5                  5                    50  avgt    3        12.000                 ms
RealisticKeyIndexImplBench.serializeModifiedIndex                                   true                  5                  5                    50  avgt    3     23629.601 ± 1442.685   us/op
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.alloc.rate                    true                  5                  5                    50  avgt    3      1883.738 ±  113.932  MB/sec
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.alloc.rate.norm               true                  5                  5                    50  avgt    3  46680665.488 ±    0.001    B/op
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.count                         true                  5                  5                    50  avgt    3         3.000             counts
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.time                          true                  5                  5                    50  avgt    3         3.000                 ms
RealisticKeyIndexImplBench.serializeUnmodifiedIndex                                 true                  5                  5                    50  avgt    3     14755.784 ± 1599.479   us/op
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.alloc.rate                  true                  5                  5                    50  avgt    3       596.915 ±   64.679  MB/sec
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.alloc.rate.norm             true                  5                  5                    50  avgt    3   9237381.971 ±    0.916    B/op
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.count                       true                  5                  5                    50  avgt    3         5.000             counts
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.time                        true                  5                  5                    50  avgt    3         3.000                 ms
```

With this change (against the 2nd commit in this PR):
```
Benchmark                                                                (deterministic)  (foldersPerLevel)  (namespaceLevels)  (tablesPerNamespace)  Mode  Cnt         Score       Error   Units
RealisticKeyIndexImplBench.deserialize                                              true                  5                  5                    50  avgt    3      9721.956 ±   899.117   us/op
RealisticKeyIndexImplBench.deserialize:·gc.alloc.rate                               true                  5                  5                    50  avgt    3      2569.404 ±   237.707  MB/sec
RealisticKeyIndexImplBench.deserialize:·gc.alloc.rate.norm                          true                  5                  5                    50  avgt    3  26197459.948 ±     0.401    B/op
RealisticKeyIndexImplBench.deserialize:·gc.count                                    true                  5                  5                    50  avgt    3         3.000              counts
RealisticKeyIndexImplBench.deserialize:·gc.time                                     true                  5                  5                    50  avgt    3         4.000                  ms
RealisticKeyIndexImplBench.deserializeAdd                                           true                  5                  5                    50  avgt    3     10992.586 ±  5473.798   us/op
RealisticKeyIndexImplBench.deserializeAdd:·gc.alloc.rate                            true                  5                  5                    50  avgt    3      2355.529 ±  1156.122  MB/sec
RealisticKeyIndexImplBench.deserializeAdd:·gc.alloc.rate.norm                       true                  5                  5                    50  avgt    3  27142468.453 ±     2.077    B/op
RealisticKeyIndexImplBench.deserializeAdd:·gc.count                                 true                  5                  5                    50  avgt    3         4.000              counts
RealisticKeyIndexImplBench.deserializeAdd:·gc.time                                  true                  5                  5                    50  avgt    3         7.000                  ms
RealisticKeyIndexImplBench.deserializeAddSerialize                                  true                  5                  5                    50  avgt    3     23902.926 ±  7150.716   us/op
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.alloc.rate                   true                  5                  5                    50  avgt    3      1451.382 ±   435.004  MB/sec
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.alloc.rate.norm              true                  5                  5                    50  avgt    3  36377289.639 ±     2.380    B/op
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.count                        true                  5                  5                    50  avgt    3         2.000              counts
RealisticKeyIndexImplBench.deserializeAddSerialize:·gc.time                         true                  5                  5                    50  avgt    3         4.000                  ms
RealisticKeyIndexImplBench.deserializeGetRandomKey                                  true                  5                  5                    50  avgt    3     10372.860 ±  5291.237   us/op
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.alloc.rate                   true                  5                  5                    50  avgt    3      2409.438 ±  1248.818  MB/sec
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.alloc.rate.norm              true                  5                  5                    50  avgt    3  26197484.208 ±     2.067    B/op
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.count                        true                  5                  5                    50  avgt    3         4.000              counts
RealisticKeyIndexImplBench.deserializeGetRandomKey:·gc.time                         true                  5                  5                    50  avgt    3         5.000                  ms
RealisticKeyIndexImplBench.deserializeIterate250                                    true                  5                  5                    50  avgt    3      9896.822 ±   362.053   us/op
RealisticKeyIndexImplBench.deserializeIterate250:·gc.alloc.rate                     true                  5                  5                    50  avgt    3      2523.990 ±    91.808  MB/sec
RealisticKeyIndexImplBench.deserializeIterate250:·gc.alloc.rate.norm                true                  5                  5                    50  avgt    3  26197492.013 ±     0.417    B/op
RealisticKeyIndexImplBench.deserializeIterate250:·gc.count                          true                  5                  5                    50  avgt    3         2.000              counts
RealisticKeyIndexImplBench.deserializeIterate250:·gc.time                           true                  5                  5                    50  avgt    3         4.000                  ms
RealisticKeyIndexImplBench.serializeModifiedIndex                                   true                  5                  5                    50  avgt    3     22920.428 ± 10699.575   us/op
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.alloc.rate                    true                  5                  5                    50  avgt    3      1474.592 ±   679.999  MB/sec
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.alloc.rate.norm               true                  5                  5                    50  avgt    3  35430641.207 ±     4.442    B/op
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.count                         true                  5                  5                    50  avgt    3         2.000              counts
RealisticKeyIndexImplBench.serializeModifiedIndex:·gc.time                          true                  5                  5                    50  avgt    3         9.000                  ms
RealisticKeyIndexImplBench.serializeUnmodifiedIndex                                 true                  5                  5                    50  avgt    3         0.007 ±     0.001   us/op
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.alloc.rate                  true                  5                  5                    50  avgt    3     11286.471 ±  1088.613  MB/sec
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.alloc.rate.norm             true                  5                  5                    50  avgt    3        80.000 ±     0.001    B/op
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.count                       true                  5                  5                    50  avgt    3        12.000              counts
RealisticKeyIndexImplBench.serializeUnmodifiedIndex:·gc.time                        true                  5                  5                    50  avgt    3         8.000                  ms
```

The benchmark was run using Java 20.0.1, using the following command.
```bash
./gradlew :nessie-versioned-storage-common:jmhJar && java -jar versioned/storage/common/build/libs/nessie-versioned-storage-common-0.62.2-SNAPSHOT-jmh.jar \
  -p namespaceLevels=5 \
  -p foldersPerLevel=5 \
  -p tablesPerNamespace=50 \
  -p deterministic=true \
  -prof gc \
  RealisticKeyIndexImplBench
```